+++ packages/react-client/src/hooks/useScreenShare.ts
@@ -1,7 +1,8 @@
 import { useCallback, useEffect, useRef } from "react";
 import { getRemoteOrLocalTrack } from "../utils/track";
-import type { ScreenshareApi, TracksMiddleware } from "../types";
+import type { ScreenshareApi } from "../types/internal";
 import { useFishjamContext } from "./useFishjamContext";
+import type { TracksMiddleware } from "../types/public";
 
 const getTracks = (stream: MediaStream): [MediaStreamTrack, MediaStreamTrack | null] => {
   const video = stream.getVideoTracks()[0];
