+++ packages/react-client/src/index.ts
@@ -13,24 +13,7 @@ export {
 } from "./hooks/public";
 export { FishjamProvider } from "./fishjamProvider";
 
-export type { PeerState, Track, PeerId, TrackId, TrackWithOrigin, Origin, PeerStatus } from "./state.types";
-
-export type {
-  DeviceManagerConfig,
-  StorageConfig,
-  Devices,
-  Device,
-  AudioDevice,
-  UseSetupMediaResult,
-  UseSetupMediaConfig,
-  ScreenshareApi,
-  UseConnect,
-  ConnectConfig,
-  TrackMiddleware,
-  PeerStateWithTracks,
-  TrackMetadata, // only for compatibility reasons, will be removed in: FCE-415
-  PeerMetadata, // only for compatibility reasons, will be removed in: FCE-415
-} from "./types";
+export * from "./types/public";
 
 export { AUDIO_TRACK_CONSTRAINTS, VIDEO_TRACK_CONSTRAINTS, SCREEN_SHARING_MEDIA_CONSTRAINTS } from "./constraints";
